requirements.txt
Django==4.1.1
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-crispy-bootstrap5==0.7
django-phonenumber-field[phonenumberslite]==7.0.0
django-js-asset==2.0.0
django-bootstrap-datepicker-plus==5.0.3


